put extractRGBA in different file in oncoprint, and fix GradientRuleSet color range
